# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Terramaps API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/maps": {
      "get": {
        "tags": [
          "Maps"
        ],
        "summary": "List Maps",
        "description": "List maps for the current user, each with its latest active job if any.",
        "operationId": "list_maps_maps_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Map"
                  },
                  "type": "array",
                  "title": "Response List Maps Maps Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Maps"
        ],
        "summary": "Create Map",
        "description": "Create map.\n\nSynchronously inserts all nodes and zip assignments, then enqueues a\nbackground task to compute geometry and data aggregations.  Returns 202\nwith the new map and an ``active_job`` tracking the pending computation.\n\nTODO:\n    - Return errors for zip codes we don't have data for\n    - Validate no duplicate layer names\n    - Validate no duplicate parents and raise error if so",
        "operationId": "create_map_maps_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportMap"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maps/{map_id}": {
      "get": {
        "tags": [
          "Maps"
        ],
        "summary": "Get Map",
        "description": "Get a single map with its latest active job.",
        "operationId": "get_map_maps__map_id__get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataFieldConfig": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number"
            ],
            "title": "Type"
          },
          "aggregations": {
            "items": {
              "type": "string",
              "enum": [
                "sum",
                "avg"
              ]
            },
            "type": "array",
            "title": "Aggregations"
          }
        },
        "type": "object",
        "required": [
          "field",
          "type",
          "aggregations"
        ],
        "title": "DataFieldConfig",
        "description": "Data field config entry stored on a map."
      },
      "DataFieldSetup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "header": {
            "type": "string",
            "title": "Header"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number"
            ],
            "title": "Type"
          },
          "aggregations": {
            "items": {
              "type": "string",
              "enum": [
                "sum",
                "avg"
              ]
            },
            "type": "array",
            "title": "Aggregations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "header",
          "type"
        ],
        "title": "DataFieldSetup",
        "description": "DataFieldSetup."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportMap": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerSetup"
            },
            "type": "array",
            "title": "Layers"
          },
          "data_fields": {
            "items": {
              "$ref": "#/components/schemas/DataFieldSetup"
            },
            "type": "array",
            "title": "Data Fields"
          },
          "values": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "type": "array",
            "title": "Values"
          },
          "headers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Headers"
          }
        },
        "type": "object",
        "required": [
          "name",
          "layers",
          "data_fields",
          "values",
          "headers"
        ],
        "title": "ImportMap",
        "description": "ImportMap."
      },
...
      "LayerSetup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "header": {
            "type": "string",
            "title": "Header"
          }
        },
        "type": "object",
        "required": [
          "name",
          "header"
        ],
        "title": "LayerSetup",
        "description": "LayerSetup."
      },
...
      "Map": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tile_version": {
            "type": "integer",
            "title": "Tile Version",
            "default": 0
          },
          "data_field_config": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DataFieldConfig"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Field Config"
          },
          "active_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MapJob"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Map",
        "description": "Map."
      },
      "MapJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_type": {
            "type": "string",
            "enum": [
              "import",
              "recompute_geometry",
              "recompute_data"
            ],
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "complete",
              "failed"
            ],
            "title": "Status"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status"
        ],
        "title": "MapJob",
        "description": "Background job for a map."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}